% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae_save.R
\name{mae_save}
\alias{mae_save}
\title{Arrange alignment and save MAE ggplot chart}
\usage{
mae_save(
  x,
  source_name,
  plot_name,
  width_pixels = 640,
  height_pixels = 450,
  logo_image_path = file.path(system.file("data", package = "maecharts"),
    "dh-logo.png")
)
}
\arguments{
\item{x}{The name of the plot you want to export to rsconnect}

\item{source_name}{The text you want to come after the text 'Source:' in the bottom left hand side of your side}

\item{plot_name}{the name of your plot - try to use as few characters as possible as rsconnect allows a maximum of 64 characters for pins}

\item{width_pixels}{Width in pixels that you want to save your chart to - defaults to 640}

\item{height_pixels}{Height in pixels that you want to save your chart to - defaults to 450}

\item{logo_image_path}{File path for the logo image you want to use in the right hand side of your chart,
which needs to be a PNG file - defaults to DH logo image that sits within the data folder of your package}
}
\value{
(Invisibly) an updated ggplot object.
}
\description{
Running this function will save your plot with the correct guidelines for publication.
It will left align your title, subtitle and source, add the MAE blocks at the bottom right and save it to your specified location.
}
\examples{
finalise_plot(x = myplot,
source = "The source for my data",
plot_name = "rsconnect_plot",
width_pixels = 640,
height_pixels = 450,
logo_image_path = "logo_image_filepath.png"
)

}
\keyword{finalise_plot}
