% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_align_plot.R, R/mae_save.R,
%   R/mae_save_local.R
\name{mae_save_local}
\alias{mae_save_local}
\title{Arrange alignment and save MAE ggplot chart}
\usage{
mae_save_local(
  x,
  source_name,
  save_file_path,
  width_pixels = 640,
  height_pixels = 450,
  logo_image_path = file.path(system.file("extdata", package = "maecharts"),
    "dh-logo.png")
)

mae_save_local(
  x,
  source_name,
  save_file_path,
  width_pixels = 640,
  height_pixels = 450,
  logo_image_path = file.path(system.file("extdata", package = "maecharts"),
    "dh-logo.png")
)

mae_save_local(
  x,
  source_name,
  save_file_path,
  width_pixels = 640,
  height_pixels = 450,
  logo_image_path = file.path(system.file("extdata", package = "maecharts"),
    "dh-logo.png")
)
}
\arguments{
\item{x}{The name of the plot you want to export to rsconnect}

\item{source_name}{The text you want to place in the bottom left hand side of your side of your chart add "Source: to the start of your string to specify a source"}

\item{save_file_path}{The name and location of the chart you want to save}

\item{width_pixels}{Width in pixels that you want to save your chart to - defaults to 640}

\item{height_pixels}{Height in pixels that you want to save your chart to - defaults to 450}

\item{logo_image_path}{File path for the logo image you want to use in the right hand side of your chart,
which needs to be a PNG file - defaults to DH logo image that sits within the data folder of your package}
}
\value{
(Invisibly) an updated ggplot object.

(Invisibly) an updated ggplot object.

(Invisibly) an updated ggplot object.
}
\description{
Running this function will save your plot with the correct guidelines for publication.
It will left align your title, subtitle and source, add the MAE blocks at the bottom right.
This function will save your plot as a pin on rsconnect which will then be automatically included in the chart_upload_table rmd where you can easily download your chart outside of the VM.

Running this function will save your plot with the correct guidelines for publication.
It will left align your title, subtitle and source, add the MAE blocks at the bottom right.
This function will save your plot as a pin on rsconnect which will then be automatically included in the chart_upload_table rmd where you can easily download your chart outside of the VM.

Running this function will save your plot with the correct guidelines for publication.
It will left align your title, subtitle and source, add the MAE blocks at the bottom right.
This function will save your plot as a pin on rsconnect which will then be automatically included in the chart_upload_table rmd where you can easily download your chart outside of the VM.
}
\keyword{mae_save_local}
