% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_mae.R
\name{scale_fill_mae}
\alias{scale_fill_mae}
\title{MAE colour scales for charting}
\usage{
scale_fill_mae(
  num_highlight_colours = 2.
  direction = T,
  type      = "categorical"
)
}
\arguments{
\item{num_highlight_colours}{The number of colours to highlight in your chart (categorical data only). The maximum is 3 with additional colours defaulting to grey.}

\item{direction}{Use argument direction = F to reverse the colour palette.}

\item{type}{The type of data being charted. Three options are available: "categorical", "sequential" and "contrast".}

\item{...}{Other arguments passed on to discrete_scale(), continuous_scale(), or binned_scale to control name, limits, breaks, labels and so forth.}
}
\description{
This argument adds the MAE colour palette to ggplot charts. The MAE colour palette is designed to be perceived by viewers with common forms of colour blindness.
}
\examples{
\dontrun{
Use categorical for distinct categories

Use sequential for continuous data
diamonds \%>\%
  ggplot() +
  geom_point(aes(x = carat, y =  price, colour = clarity)) +
  scale_colour_mae(type = "sequential")

Use contrast for data with two distinct categories. Contrast supports 6 shades of two contrasting colours
Titanic \%>\%
  as_tibble() \%>\%
  group_by(Class, Survived, Sex) \%>\%
  summarise(n = sum(n)) \%>\%
  ungroup() \%>\%
  mutate(legend = paste(Survived, Sex),
         legend = factor(legend, levels = c("No Female", "Yes Female", "No Male", "Yes Male"), labels = c("No ", "Yes ", "No", "Yes"))) \%>\%
  ggplot() +
  geom_col(aes(x = Class, y = n, fill = legend)) +
  facet_wrap(~Sex) +
  scale_fill_mae(type = "contrast", direction = F) +
  theme_mae() +
  labs(title = "Survivors of the titanic", subtitle = "By sex and passenger class",
       x = "", y = "")
}
}
