% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mae.R
\name{theme_mae}
\alias{theme_mae}
\title{Add mae theme to ggplot chart}
\usage{
theme_mae()
}
\description{
This function allows you to add the mae theme to your ggplot graphics. Individual theme elements can be overwritten in subsequent ggplot code.
}
\examples{
iris \%>\%
 ggplot() +
 geom_point(aes(x = Sepal.Length, y = Sepal.Width, colour = Species)) +
 theme_mae()

MAE theme elements can be overwritten by adding additional
theme arguments after theme_mae() call

iris \%>\%
 ggplot() +
 geom_point(aes(x = Sepal.Length, y = Sepal.Width, colour = Species)) +
 theme_mae() +
 theme(legend.position = 'right')
}
\keyword{theme_mae}
